rename BaseNavigationActivity.kt to HomeActivity.kt
