+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
         android:theme="@style/Theme.WeatherApp"
         tools:targetApi="31">
         <activity
-            android:name=".feature.base.BaseNavigationActivity"
+            android:name=".feature.home.HomeActivity"
             android:exported="true"
             android:theme="@style/Theme.WeatherApp">
             <intent-filter>
